refactor api v1 routers and add support for redoc docs
