+++ etc/motley_cue.conf
@@ -14,6 +14,8 @@ log_level = WARNING
 # enable_docs = False
 ## location of swagger docs -- default: /docs
 # docs_url = /docs
+## location of redoc docs -- default: /redoc
+# redoc_url = /redoc
 
 ############
 [mapper.otp]
