+++ motley_cue/api/__init__.py
@@ -2,7 +2,7 @@ from fastapi import FastAPI
 from fastapi.exceptions import RequestValidationError, ResponseValidationError
 from pydantic import ValidationError
 
-from motley_cue.api.api_v1.api import api_router as api_router_v1
+from motley_cue.api.api_v1 import api_router as api_router_v1
 from motley_cue.dependencies import mapper, Settings
 from motley_cue.mapper.exceptions import (
     validation_exception_handler,
@@ -13,7 +13,7 @@ from motley_cue.mapper.exceptions import (
 def create_app():
     """Create motley_cue api."""
 
-    settings = Settings(docs_url=mapper.config.docs_url)
+    settings = Settings(docs_url=mapper.config.docs_url, redoc_url=mapper.config.redoc_url)
     app = FastAPI(
         title=settings.title,
         description=settings.description,
@@ -29,11 +29,12 @@ def create_app():
     app.add_exception_handler(ValidationError, validation_exception_handler)
     app.add_exception_handler(ResponseValidationError, validation_exception_handler)
 
+    # for compatibility with old API, include all endpoints in the root
     app.include_router(api_router_v1, prefix="", tags=["API"])
+    # latest API version
+    app.include_router(api_router_v1, prefix=settings.API_LATEST_STR, tags=["API latest"])
+    # all API versions
     app.include_router(api_router_v1, prefix=settings.API_V1_STR, tags=["API v1"])
-    app.include_router(
-        api_router_v1, prefix=settings.API_LATEST_STR, tags=["API latest"]
-    )
 
     return app
